refactor: Move script in its own module
The main reason is that the fkie_cve_make_release package
imports `Feed` in its module initializer.
This leads to repo.py being imported which in turn raises
a ValueError if the nvd-release repository is not on disk.
By moving the cli in its own package we can both keep the imports
and print the help page before the error is raised.